Fix Battlemon progression through level 250
Amp-Thread-ID: https://ampcode.com/threads/T-01a00000-fda1-71df-9051-7e6cb6f95bed

+++ entities/src/main/java/com/jftse/entities/database/model/pet/Pet.java
@@ -24,7 +24,7 @@ public class Pet extends AbstractBaseModel {
     private Byte type;
     private String name;
     @Column(columnDefinition = "TINYINT UNSIGNED")
-    private Byte level;
+    private Integer level;
     private Integer expPoints;
     private Integer hp;
     private Byte strength;
