+++ game-server/src/main/java/com/jftse/emulator/server/core/packets/pet/S2CPetDataAnswerPacket.java
@@ -17,7 +17,7 @@ public class S2CPetDataAnswerPacket extends Packet {
                         this.write(
                                 pet.getType(),
                                 pet.getName(),
-                                pet.getLevel(),
+                                pet.getLevel().byteValue(),
                                 pet.getExpPoints(),
                                 pet.getHp(),
                                 pet.getStrength(),
